made _mult_fixed_window the default point multiplication
