+++ btclib/tests/test_curvegroupf.py
@@ -12,7 +12,7 @@
 
 import pytest
 
-from btclib.curve import CurveGroup, _mult_aff
+from btclib.curvegroup import CurveGroup, _mult_aff
 from btclib.curvegroupf import find_all_points, find_subgroup_points
