+++ btclib/tests/test_curvemult.py
@@ -15,33 +15,182 @@ import secrets
 import pytest
 
 from btclib.alias import INF, INFJ
-from btclib.curvegroup import _jac_from_aff, _mult_jac
-from btclib.curvemult import (
+from btclib.curvegroup import (
     _double_mult,
+    _jac_from_aff,
+    _mult_aff,
+    _mult_base_3,
+    _mult_fixed_window,
+    _mult_jac,
+    _mult_mont_ladder,
     _multi_mult,
-    double_mult,
-    mult,
-    multi_mult,
+    multiples,
 )
+from btclib.curvemult import double_mult, mult, multi_mult
 from btclib.curves import secp256k1
 from btclib.pedersen import second_generator
-from btclib.tests.test_curves import low_card_curves
+from btclib.tests.test_curves import all_curves, low_card_curves
 
 ec23_31 = low_card_curves["ec23_31"]
 
 
+@pytest.mark.second
+def test_mult_aff() -> None:
+    for ec in all_curves.values():
+        assert _mult_aff(0, ec.G, ec) == INF
+        assert _mult_aff(0, INF, ec) == INF
+
+        assert _mult_aff(1, INF, ec) == INF
+        assert _mult_aff(1, ec.G, ec) == ec.G
+
+        P = ec._add_aff(ec.G, ec.G)
+        assert P == _mult_aff(2, ec.G, ec)
+
+        P = _mult_aff(ec.n - 1, ec.G, ec)
+        assert ec.negate(ec.G) == P
+        assert _mult_aff(ec.n - 1, INF, ec) == INF
+
+        assert ec._add_aff(P, ec.G) == INF
+        assert _mult_aff(ec.n, ec.G, ec) == INF
+        assert _mult_aff(ec.n, INF, ec) == INF
+
+        with pytest.raises(ValueError, match="negative m: -0x"):
+            _mult_aff(-1, ec.G, ec)
+
+
+def test_mult_jac() -> None:
+    for ec in all_curves.values():
+        assert ec._jac_equality(_mult_jac(0, ec.GJ, ec), INFJ)
+        assert ec._jac_equality(_mult_jac(0, INFJ, ec), INFJ)
+
+        assert ec._jac_equality(_mult_jac(1, INFJ, ec), INFJ)
+        assert ec._jac_equality(_mult_jac(1, ec.GJ, ec), ec.GJ)
+
+        PJ = ec._add_jac(ec.GJ, ec.GJ)
+        assert ec._jac_equality(PJ, _mult_jac(2, ec.GJ, ec))
+
+        PJ = _mult_jac(ec.n - 1, ec.GJ, ec)
+        assert ec._jac_equality(ec.negate_jac(ec.GJ), PJ)
+
+        assert ec._jac_equality(_mult_jac(ec.n - 1, INFJ, ec), INFJ)
+        assert ec._jac_equality(ec._add_jac(PJ, ec.GJ), INFJ)
+        assert ec._jac_equality(_mult_jac(ec.n, ec.GJ, ec), INFJ)
+
+        with pytest.raises(ValueError, match="negative m: -0x"):
+            _mult_jac(-1, ec.GJ, ec)
+
+
+def test_mult() -> None:
+    for ec in low_card_curves.values():
+        for q in range(ec.n):
+            Q = _mult_aff(q, ec.G, ec)
+            assert ec.is_on_curve(Q), f"{q}, {ec}"
+            QJ = _mult_fixed_window(q, ec.GJ, ec)
+            assert ec.is_on_curve(ec._aff_from_jac(QJ)), f"{q}, {ec}"
+            assert Q == ec._aff_from_jac(QJ), f"{q}, {ec}"
+        assert INF == _mult_aff(q, INF, ec), f"{q}, {ec}"
+        assert ec._jac_equality(INFJ, _mult_fixed_window(q, INFJ, ec)), f"{q}, {ec}"
+
+
+def test_mont_ladder() -> None:
+    for ec in low_card_curves.values():
+        assert ec._jac_equality(_mult_mont_ladder(0, ec.GJ, ec), INFJ)
+        assert ec._jac_equality(_mult_mont_ladder(0, INFJ, ec), INFJ)
+
+        assert ec._jac_equality(_mult_mont_ladder(1, INFJ, ec), INFJ)
+        assert ec._jac_equality(_mult_mont_ladder(1, ec.GJ, ec), ec.GJ)
+
+        PJ = _mult_mont_ladder(2, ec.GJ, ec)
+        assert ec._jac_equality(PJ, ec._add_jac(ec.GJ, ec.GJ))
+
+        PJ = _mult_mont_ladder(ec.n - 1, ec.GJ, ec)
+        assert ec._jac_equality(ec.negate_jac(ec.GJ), PJ)
+
+        assert ec._jac_equality(_mult_mont_ladder(ec.n - 1, INFJ, ec), INFJ)
+        assert ec._jac_equality(ec._add_jac(PJ, ec.GJ), INFJ)
+        assert ec._jac_equality(_mult_mont_ladder(ec.n, ec.GJ, ec), INFJ)
+
+        with pytest.raises(ValueError, match="negative m: "):
+            _mult_mont_ladder(-1, ec.GJ, ec)
+
+    ec = ec23_31
+    for k1 in range(ec.n):
+        K1 = _mult_mont_ladder(k1, ec.GJ, ec)
+        assert ec._jac_equality(K1, _mult_fixed_window(k1, ec.GJ, ec))
+
+
+def test_mult_base_3() -> None:
+    for ec in low_card_curves.values():
+        assert ec._jac_equality(_mult_base_3(0, ec.GJ, ec), INFJ)
+        assert ec._jac_equality(_mult_base_3(0, INFJ, ec), INFJ)
+
+        assert ec._jac_equality(_mult_base_3(1, INFJ, ec), INFJ)
+        assert ec._jac_equality(_mult_base_3(1, ec.GJ, ec), ec.GJ)
+
+        PJ = _mult_base_3(2, ec.GJ, ec)
+        assert ec._jac_equality(PJ, ec._add_jac(ec.GJ, ec.GJ))
+
+        PJ = _mult_base_3(ec.n - 1, ec.GJ, ec)
+        assert ec._jac_equality(ec.negate_jac(ec.GJ), PJ)
+
+        assert ec._jac_equality(_mult_base_3(ec.n - 1, INFJ, ec), INFJ)
+        assert ec._jac_equality(ec._add_jac(PJ, ec.GJ), INFJ)
+        assert ec._jac_equality(_mult_base_3(ec.n, ec.GJ, ec), INFJ)
+
+        with pytest.raises(ValueError, match="negative m: "):
+            _mult_base_3(-1, ec.GJ, ec)
+
+    ec = ec23_31
+    for k1 in range(ec.n):
+        K1 = _mult_base_3(k1, ec.GJ, ec)
+        assert ec._jac_equality(K1, _mult_fixed_window(k1, ec.GJ, ec))
+
+
+def test_mult_fixed_window() -> None:
+    for w in range(1, 10):  # Actually it makes use of w=4 or w=5, only to check
+        for ec in low_card_curves.values():
+            assert ec._jac_equality(_mult_fixed_window(0, ec.GJ, ec, w), INFJ)
+            assert ec._jac_equality(_mult_fixed_window(0, INFJ, ec, w), INFJ)
+
+            assert ec._jac_equality(_mult_fixed_window(1, INFJ, ec, w), INFJ)
+            assert ec._jac_equality(_mult_fixed_window(1, ec.GJ, ec, w), ec.GJ)
+
+            PJ = _mult_fixed_window(2, ec.GJ, ec, w)
+            assert ec._jac_equality(PJ, ec._add_jac(ec.GJ, ec.GJ))
+
+            PJ = _mult_fixed_window(ec.n - 1, ec.GJ, ec, w)
+            assert ec._jac_equality(ec.negate_jac(ec.GJ), PJ)
+
+            assert ec._jac_equality(_mult_fixed_window(ec.n - 1, INFJ, ec, w), INFJ)
+            assert ec._jac_equality(ec._add_jac(PJ, ec.GJ), INFJ)
+            assert ec._jac_equality(_mult_fixed_window(ec.n, ec.GJ, ec, w), INFJ)
+
+            with pytest.raises(ValueError, match="negative m: "):
+                _mult_fixed_window(-1, ec.GJ, ec, w)
+
+            with pytest.raises(ValueError, match="non positive w: "):
+                _mult_fixed_window(1, ec.GJ, ec, -w)
+
+    ec = ec23_31
+    for w in range(1, 10):
+        for k1 in range(ec.n):
+            K1 = _mult_fixed_window(k1, ec.GJ, ec, w)
+            assert ec._jac_equality(K1, _mult_jac(k1, ec.GJ, ec))
+
+
+@pytest.mark.fifth
 def test_assorted_mult() -> None:
     ec = ec23_31
     H = second_generator(ec)
     HJ = _jac_from_aff(H)
     for k1 in range(ec.n):
-        K1J = _mult_jac(k1, ec.GJ, ec)
+        K1J = _mult_fixed_window(k1, ec.GJ, ec)
         for k2 in range(ec.n):
-            K2J = _mult_jac(k2, HJ, ec)
+            K2J = _mult_fixed_window(k2, HJ, ec)
 
             shamir = _double_mult(k1, ec.GJ, k2, ec.GJ, ec)
             assert ec.is_on_curve(ec._aff_from_jac(shamir))
-            assert ec._jac_equality(shamir, _mult_jac(k1 + k2, ec.GJ, ec))
+            assert ec._jac_equality(shamir, _mult_fixed_window(k1 + k2, ec.GJ, ec))
 
             shamir = _double_mult(k1, INFJ, k2, HJ, ec)
             assert ec.is_on_curve(ec._aff_from_jac(shamir))
@@ -57,7 +206,7 @@ def test_assorted_mult() -> None:
             assert ec._jac_equality(K1JK2J, shamir)
 
             k3 = 1 + secrets.randbelow(ec.n - 1)
-            K3J = _mult_jac(k3, ec.GJ, ec)
+            K3J = _mult_fixed_window(k3, ec.GJ, ec)
             K1JK2JK3J = ec._add_jac(K1JK2J, K3J)
             assert ec.is_on_curve(ec._aff_from_jac(K1JK2JK3J))
             boscoster = _multi_mult([k1, k2, k3], [ec.GJ, HJ, ec.GJ], ec)
@@ -66,7 +215,7 @@ def test_assorted_mult() -> None:
             assert ec._jac_equality(K1JK2JK3J, boscoster)
 
             k4 = 1 + secrets.randbelow(ec.n - 1)
-            K4J = _mult_jac(k4, HJ, ec)
+            K4J = _mult_fixed_window(k4, HJ, ec)
             K1JK2JK3JK4J = ec._add_jac(K1JK2JK3J, K4J)
             assert ec.is_on_curve(ec._aff_from_jac(K1JK2JK3JK4J))
             points = [ec.GJ, HJ, ec.GJ, HJ]
@@ -93,6 +242,7 @@ def test_assorted_mult() -> None:
         _double_mult(1, HJ, -5, ec.GJ, ec)
 
 
+@pytest.mark.fourth
 def test_assorted_mult2() -> None:
     ec = ec23_31
     H = second_generator(ec)
@@ -208,3 +358,55 @@ def test_double_mult() -> None:
     # FIXME
     # T = multi_mult([-5, 1], [H, G])
     # assert T == exp
+
+
+def test_multiples() -> None:
+
+    ec = secp256k1
+    with pytest.raises(ValueError, match="size too low: "):
+        multiples(ec.GJ, 1, ec)
+
+    T = [INFJ, ec.GJ]
+    M = multiples(ec.GJ, 2, ec)
+    assert len(M) == 2
+    assert M == T
+
+    T.append(ec._double_jac(ec.GJ))
+    M = multiples(ec.GJ, 3, ec)
+    assert len(M) == 3
+    assert M == T
+
+    T.append(ec._add_jac(T[-1], ec.GJ))
+    M = multiples(ec.GJ, 4, ec)
+    assert len(M) == 4
+    assert M == T
+
+    T.append(ec._double_jac(T[2]))
+    M = multiples(ec.GJ, 5, ec)
+    assert len(M) == 5
+    assert M == T
+
+    T.append(ec._add_jac(T[-1], ec.GJ))
+    M = multiples(ec.GJ, 6, ec)
+    assert len(M) == 6
+    assert M == T
+
+    T.append(ec._double_jac(T[3]))
+    M = multiples(ec.GJ, 7, ec)
+    assert len(M) == 7
+    assert M == T
+
+    T.append(ec._add_jac(T[-1], ec.GJ))
+    M = multiples(ec.GJ, 8, ec)
+    assert len(M) == 8
+    assert M == T
+
+    T.append(ec._double_jac(T[4]))
+    M = multiples(ec.GJ, 9, ec)
+    assert len(M) == 9
+    assert M == T
+
+    T.append(ec._add_jac(T[-1], ec.GJ))
+    M = multiples(ec.GJ, 10, ec)
+    assert len(M) == 10
+    assert M == T
