+++ btclib/tests/test_dsa.py
@@ -16,7 +16,7 @@ import pytest
 
 from btclib import dsa
 from btclib.alias import INF
-from btclib.curvemult import _mult_jac, double_mult, mult
+from btclib.curvemult import _mult_fixed_window, double_mult, mult
 from btclib.curves import CURVES
 from btclib.numbertheory import mod_inv
 from btclib.rfc6979 import rfc6979
@@ -160,9 +160,9 @@ def test_low_cardinality() -> None:
     # only low cardinality test curves or it would take forever
     for ec in test_curves:
         for q in range(1, ec.n):  # all possible private keys
-            QJ = _mult_jac(q, ec.GJ, ec)  # public key
+            QJ = _mult_fixed_window(q, ec.GJ, ec)  # public key
             for k in range(1, ec.n):  # all possible ephemeral keys
-                RJ = _mult_jac(k, ec.GJ, ec)
+                RJ = _mult_fixed_window(k, ec.GJ, ec)
                 r = ec._x_aff_from_jac(RJ) % ec.n
                 k_inv = mod_inv(k, ec.n)
                 for e in range(ec.n):  # all possible challenges
