+++ btclib/tests/test_rfc6979.py
@@ -46,7 +46,7 @@ def test_rfc6979_example() -> None:
     assert k == rfc6979(msg, x, fake_ec)  # type: ignore
 
 
-@pytest.mark.second
+@pytest.mark.third
 def test_rfc6979_tv() -> None:
 
     fname = "rfc6979.json"
