+++ setup.cfg
@@ -1,5 +1,5 @@
 [tool:pytest]
-addopts = --durations=8 -n auto
+addopts = --durations=12 -n auto
 # Longest test should go first, to favor running tests in parallel
 markers =
     first: ordering using pytest-ordering
@@ -8,5 +8,3 @@ markers =
     fourth: ordering using pytest-ordering
     fifth: ordering using pytest-ordering
     sixth: ordering using pytest-ordering
-    seventh: ordering using pytest-ordering
-    eighth: ordering using pytest-ordering
